# api/openapi/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger Leasing-car  - OpenAPI 3.0",
    "version": "1.0.0",
    "description": "Swagger for Leasing-car project\n\n[Github](https://github.com/romainwg/leasing-car)"
  },
  "servers": [
    {
      "url": "https://api.leasing-car.r-wg.it"
    }
  ],
  "tags": [
    {
      "name": "customer"
    }
  ],
  "paths": {
    "/customer/get/{customer_id}": {
      "get": {
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "description": "ID of customer",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "tags": [
          "customer"
        ],
        "summary": "get-customer",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "invalid ID supplied"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/customer/getall": {
      "get": {
        "tags": [
          "customer"
        ],
        "summary": "get-all-customers",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/customer/create": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "add-customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "test@contact.com",
                  "name": "Nametest'",
                  "firstname": "Firstname",
                  "birthday": "1950-06-20T00:00:00Z",
                  "driving_licence_number": "TESTT654051SM9IJ"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "405": {
            "description": "invalid input"
          },
          "406": {
            "description": "email already exists"
          },
          "407": {
            "description": "driving licence number already exists"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/customer/update/{customer_id}": {
      "put": {
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "description": "ID of customer",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "tags": [
          "customer"
        ],
        "summary": "update-customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "contact@contact.com",
                  "name": "Name",
                  "firstname": "Firstname",
                  "birthday": "1985-06-20T00:00:00Z",
                  "driving_licence_number": "MORGA657054SM9IJ"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": 2,
                    "email": "contact@contact.com",
                    "name": "Name",
                    "firstname": "Firstname",
                    "birthday": "1985-06-20T00:00:00Z",
                    "driving_licence_number": "MORGA657054SM9IJ",
                    "cars": {
                      "type": "array",
                      "items": {
                        "id": 2
                      },
                      "email": "contact@contact.com",
                      "name": "Name",
                      "firstname": "Firstname",
                      "birthday": "1985-06-20T00:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "405": {
            "description": "invalid ID supplied"
          },
          "406": {
            "description": "email already exists"
          },
          "407": {
            "description": "driving licence number already exists"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/customer-car/associate": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "associate-customer2car",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "customer_id": 2,
                  "car_id": 3
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "invalid ID supplied"
          },
          "405": {
            "description": "invalid input"
          },
          "406": {
            "description": "car already associated"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/customer-car/disassociate": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "disassociate-customer2car",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "customer_id": 4,
                  "car_id": 12
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "invalid ID supplied"
          },
          "405": {
            "description": "invalid input"
          },
          "406": {
            "description": "car already associated"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 3
          },
          "email": {
            "type": "string",
            "example": "contact@r-wg.it"
          },
          "name": {
            "type": "string",
            "example": "Romain"
          },
          "firstname": {
            "type": "string",
            "example": "WG"
          },
          "birthday": {
            "type": "string",
            "format": "date-time",
            "example": "1990-12-13T00:00:00Z"
          },
          "driving_licence_number": {
            "type": "string",
            "example": "ROMAW657054SM9IJ"
          },
          "cars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Car"
            }
          }
        }
      },
      "Car": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "matriculation_number": {
            "type": "string",
            "example": "ZA321BC"
          },
          "brand": {
            "type": "string",
            "example": "Peugeot"
          },
          "model": {
            "type": "string",
            "example": "405"
          },
          "year": {
            "type": "string",
            "format": "int32",
            "example": 1994
          }
        }
      }
    }
  }
}